// backup/graphql-types.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  /** The `ID` scalar type represents a unique identifier, often used to refetch an object or as key for a cache. The ID type appears in a JSON response as a String; however, it is not intended to be human-readable. When expected as an input type, any string (such as `"4"`) or integer (such as `4`) input value will be accepted as an ID. */
  ID: string;
  /** The `String` scalar type represents textual data, represented as UTF-8 character sequences. The String type is most often used by GraphQL to represent free-form human-readable text. */
  String: string;
  /** The `Boolean` scalar type represents `true` or `false`. */
  Boolean: boolean;
  /** The `Int` scalar type represents non-fractional signed whole numeric values. Int can represent values between -(2^31) and 2^31 - 1. */
  Int: number;
  /** The `Float` scalar type represents signed double-precision fractional values as specified by [IEEE 754](https://en.wikipedia.org/wiki/IEEE_floating_point). */
  Float: number;
  /** A date string, such as 2007-12-03, compliant with the ISO 8601 standard for representation of dates and times using the Gregorian calendar. */
  Date: any;
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: any;
};
// ...
export type Node = {
  id: Scalars['ID'];
  parent?: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
};

export type Internal = {
  content?: Maybe<Scalars['String']>;
  contentDigest: Scalars['String'];
  description?: Maybe<Scalars['String']>;
  fieldOwners?: Maybe<Array<Maybe<Scalars['String']>>>;
  ignoreType?: Maybe<Scalars['Boolean']>;
  mediaType?: Maybe<Scalars['String']>;
  owner: Scalars['String'];
  type: Scalars['String'];
};
// ...
export type Wordpress_Node = {
  /** The globally unique ID for the object */
  id: Scalars['ID'];
};
// ...
export type Wordpress_UniformResourceIdentifiable = {
  /** The unique resource identifier path */
  id: Scalars['ID'];
  /** Whether the node is a Content Node */
  isContentNode: Scalars['Boolean'];
  /** Whether the node is a Term */
  isTermNode: Scalars['Boolean'];
  /** The unique resource identifier path */
  uri?: Maybe<Scalars['String']>;
};
// ...
export type Wordpress_User = Node & Wordpress_Node & Wordpress_UniformResourceIdentifiable & Wordpress_Commenter & Wordpress_DatabaseIdentifier & {
  /** Avatar object for user. The avatar object can be retrieved in different sizes by specifying the size argument. */
  avatar?: Maybe<Wordpress_Avatar>;
  /** User metadata option name. Usually it will be &quot;wp_capabilities&quot;. */
  capKey?: Maybe<Scalars['String']>;
  /** A list of capabilities (permissions) granted to the user */
  capabilities?: Maybe<Array<Maybe<Scalars['String']>>>;
  /** Connection between the User type and the Comment type */
  comments?: Maybe<Wordpress_UserToCommentConnection>;
  /** Identifies the primary key from the database. */
  databaseId: Scalars['Int'];
  /** Description of the user. */
  description?: Maybe<Scalars['String']>;
  /** Email address of the user. This is equivalent to the WP_User-&gt;user_email property. */
  email?: Maybe<Scalars['String']>;
  /** A complete list of capabilities including capabilities inherited from a role. This is equivalent to the array keys of WP_User-&gt;allcaps. */
  extraCapabilities?: Maybe<Array<Maybe<Scalars['String']>>>;
  /** First name of the user. This is equivalent to the WP_User-&gt;user_first_name property. */
  firstName?: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  /** Whether the node is a Content Node */
  isContentNode: Scalars['Boolean'];
  /** Whether the node is a Term */
  isTermNode: Scalars['Boolean'];
  /** Last name of the user. This is equivalent to the WP_User-&gt;user_last_name property. */
  lastName?: Maybe<Scalars['String']>;
  /** The preferred language locale set for the user. Value derived from get_user_locale(). */
  locale?: Maybe<Scalars['String']>;
  /** Display name of the user. This is equivalent to the WP_User-&gt;dispaly_name property. */
  name?: Maybe<Scalars['String']>;
  /** The nicename for the user. This field is equivalent to WP_User-&gt;user_nicename */
  nicename?: Maybe<Scalars['String']>;
  /** Nickname of the user. */
  nickname?: Maybe<Scalars['String']>;
  /** Connection between the User type and the page type */
  pages?: Maybe<Wordpress_UserToPageConnection>;
  /** Connection between the User type and the post type */
  posts?: Maybe<Wordpress_UserToPostConnection>;
  /** The date the user registered or was created. The field follows a full ISO8601 date string format. */
  registeredDate?: Maybe<Scalars['String']>;
  /** Connection between the User type and the UserRole type */
  roles?: Maybe<Wordpress_UserToUserRoleConnection>;
  /** The slug for the user. This field is equivalent to WP_User-&gt;user_nicename */
  slug?: Maybe<Scalars['String']>;
  /** The unique resource identifier path */
  uri?: Maybe<Scalars['String']>;
  /** A website url that is associated with the user. */
  url?: Maybe<Scalars['String']>;
  /** Username for the user. This field is equivalent to WP_User-&gt;user_login. */
  username?: Maybe<Scalars['String']>;
  nodeType?: Maybe<Scalars['String']>;
  parent?: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
};

export type Wordpress_Commenter = {
  /** Avatar object for user. The avatar object can be retrieved in different sizes by specifying the size argument. */
  avatar?: Maybe<Wordpress_Avatar>;
  /** Identifies the primary key from the database. */
  databaseId: Scalars['Int'];
  /** The email address of the author of a comment. */
  email?: Maybe<Scalars['String']>;
  /** The globally unique identifier for the comment author. */
  id: Scalars['ID'];
  /** The name of the author of a comment. */
  name?: Maybe<Scalars['String']>;
  /** The url of the author of a comment. */
  url?: Maybe<Scalars['String']>;
};

export type Wordpress_DatabaseIdentifier = {
  /** The unique identifier stored in the database */
  databaseId: Scalars['Int'];
};

/** Avatars are profile images for users. WordPress by default uses the Gravatar service to host and fetch avatars from. */
export type Wordpress_Avatar = {
  /** URL for the default image or a default type. Accepts &#039;404&#039; (return a 404 instead of a default image), &#039;retro&#039; (8bit), &#039;monsterid&#039; (monster), &#039;wavatar&#039; (cartoon face), &#039;indenticon&#039; (the &#039;quilt&#039;), &#039;mystery&#039;, &#039;mm&#039;, or &#039;mysteryman&#039; (The Oyster Man), &#039;blank&#039; (transparent GIF), or &#039;gravatar_default&#039; (the Gravatar logo). */
  default?: Maybe<Scalars['String']>;
  /** HTML attributes to insert in the IMG element. Is not sanitized. */
  extraAttr?: Maybe<Scalars['String']>;
  /** Whether to always show the default image, never the Gravatar. */
  forceDefault?: Maybe<Scalars['Boolean']>;
  /** Whether the avatar was successfully found. */
  foundAvatar?: Maybe<Scalars['Boolean']>;
  /** Height of the avatar image. */
  height?: Maybe<Scalars['Int']>;
  /** What rating to display avatars up to. Accepts &#039;G&#039;, &#039;PG&#039;, &#039;R&#039;, &#039;X&#039;, and are judged in that order. */
  rating?: Maybe<Scalars['String']>;
  /** Type of url scheme to use. Typically HTTP vs. HTTPS. */
  scheme?: Maybe<Scalars['String']>;
  /** The size of the avatar in pixels. A value of 96 will match a 96px x 96px gravatar image. */
  size?: Maybe<Scalars['Int']>;
  /** URL for the gravatar image source. */
  url?: Maybe<Scalars['String']>;
  /** Width of the avatar image. */
  width?: Maybe<Scalars['Int']>;
};

/** Connection between the User type and the Comment type */
export type Wordpress_UserToCommentConnection = {
  /** The nodes of the connection, without the edges */
  nodes?: Maybe<Array<Maybe<Wordpress_Comment>>>;
};
// ...
export type Wordpress_UserToPageConnection = {
  /** The nodes of the connection, without the edges */
  nodes?: Maybe<Array<Maybe<Wordpress_Page>>>;
};
// ...
export type Wordpress_Post_Postacf = {
  division?: Maybe<Scalars['String']>;
  fieldGroupName?: Maybe<Scalars['String']>;
  flashPost?: Maybe<Wordpress_Post_Postacf_FlashPost>;
  matchReport?: Maybe<Wordpress_Post_Postacf_MatchReport>;
  normalPost?: Maybe<Wordpress_Post_Postacf_NormalPost>;
  postCategory?: Maybe<Scalars['String']>;
};

/** Field Group */
export type Wordpress_Post_Postacf_FlashPost = {
  end?: Maybe<Scalars['String']>;
  fieldGroupName?: Maybe<Scalars['String']>;
  start?: Maybe<Scalars['String']>;
};

/** Field Group */
export type Wordpress_Post_Postacf_MatchReport = {
  assists?: Maybe<Scalars['String']>;
  fieldGroupName?: Maybe<Scalars['String']>;
  gameDay?: Maybe<Scalars['String']>;
  penalties?: Maybe<Scalars['String']>;
  /** Die Drittelergebniss als Text eingeben. Beispiel (1:0, 0:0, 2:2, 1:0) n.P. */
  periodResults?: Maybe<Scalars['String']>;
  scoreGuest?: Maybe<Scalars['Float']>;
  scoreHome?: Maybe<Scalars['Float']>;
  scorer?: Maybe<Scalars['String']>;
  /** Wurde das Spiel in der Nachspielzeit oder nach Penalty schießen beendet? Wurde es Abgebrochen? */
  specialCase?: Maybe<Scalars['String']>;
  spectators?: Maybe<Scalars['String']>;
  teamGuest?: Maybe<Scalars['String']>;
  teamGuestSlug?: Maybe<Scalars['String']>;
  teamHome?: Maybe<Scalars['String']>;
  teamHomeSlug?: Maybe<Scalars['String']>;
};

/** Field Group */
export type Wordpress_Post_Postacf_NormalPost = {
  fieldGroupName?: Maybe<Scalars['String']>;
  /** Hier ist Platz für maximal 9 Bilder. */
  imageGallery?: Maybe<Array<Maybe<Wordpress_MediaItem>>>;
};
// ...
export type Wordpress_UserToPostConnection = {
  /** The nodes of the connection, without the edges */
  nodes?: Maybe<Array<Maybe<Wordpress_Post>>>;
};

/** Connection between the User type and the UserRole type */
export type Wordpress_UserToUserRoleConnection = {
  /** The nodes of the connection, without the edges */
  nodes?: Maybe<Array<Maybe<Wordpress_UserRole>>>;
};
// ...
export type ImageSharpFluid = {
  base64?: Maybe<Scalars['String']>;
  tracedSVG?: Maybe<Scalars['String']>;
  aspectRatio: Scalars['Float'];
  src: Scalars['String'];
  srcSet: Scalars['String'];
  srcWebp?: Maybe<Scalars['String']>;
  srcSetWebp?: Maybe<Scalars['String']>;
  sizes: Scalars['String'];
  originalImg?: Maybe<Scalars['String']>;
  originalName?: Maybe<Scalars['String']>;
  presentationWidth: Scalars['Int'];
  presentationHeight: Scalars['Int'];
};
// ...
export type BlogPostFieldsFragment = (
  Pick<Wordpress_Post, 'date' | 'content' | 'title' | 'id' | 'uri' | 'modified'>
  & { featuredImage?: Maybe<{ node?: Maybe<(
      Pick<Wordpress_MediaItem, 'altText' | 'title' | 'id'>
      & { localFile?: Maybe<{ childImageSharp?: Maybe<{ fluid?: Maybe<FluidWithWebpFieldsFragment> }> }> }
    )> }>, author?: Maybe<{ node?: Maybe<Pick<Wordpress_User, 'name' | 'nicename' | 'nickname'>> }>, postACF?: Maybe<(
    Pick<Wordpress_Post_Postacf, 'division' | 'postCategory'>
    & NormalPostFieldsFragment
    & MatchReportFieldsFragment
  )> }
);

export type MatchReportFieldsFragment = { matchReport?: Maybe<Pick<Wordpress_Post_Postacf_MatchReport, 'assists' | 'gameDay' | 'penalties' | 'scoreGuest' | 'scoreHome' | 'scorer' | 'spectators' | 'teamGuest' | 'teamGuestSlug' | 'teamHome' | 'teamHomeSlug' | 'periodResults' | 'specialCase'>> };

export type NormalPostFieldsFragment = { normalPost?: Maybe<{ imageGallery?: Maybe<Array<Maybe<(
      Pick<Wordpress_MediaItem, 'id' | 'title' | 'altText'>
      & { localFile?: Maybe<{ childImageSharp?: Maybe<{ fluid?: Maybe<FluidWithWebpFieldsFragment> }> }> }
    )>>> }> };
// ...
export type FluidWithWebpFieldsFragment = Pick<ImageSharpFluid, 'src' | 'sizes' | 'srcWebp' | 'srcSet' | 'aspectRatio'>;